Feat: team 토큰 사용 인증 인가 추가

+++ src/main/kotlin/com/arij/ajir/domain/team/service/TeamService.kt
@@ -1,9 +1,11 @@
 package com.arij.ajir.domain.team.service
 
+import com.arij.ajir.common.dto.UserProfileDto
 import com.arij.ajir.common.exception.DuplicateArgumentException
 import com.arij.ajir.common.exception.ModelNotFoundException
+import com.arij.ajir.common.exception.NotAuthorityException
+import com.arij.ajir.domain.member.model.Role
 import com.arij.ajir.domain.member.repository.MemberRepository
-import com.arij.ajir.domain.member.service.MemberService
 import com.arij.ajir.domain.team.dto.TeamRequest
 import com.arij.ajir.domain.team.dto.TeamResponse
 import com.arij.ajir.domain.team.model.Team
@@ -19,28 +21,26 @@ class TeamService(
     private val memberRepository: MemberRepository
 ) {
 
-    fun createTeams(teamRequest: TeamRequest, /*memberId : Long*/): TeamResponse {
-        //TODO("Team Repository 에서 teamRequest.name 과 같은 이름이 있을 경우 throw DuplicateArgumentException")
+    fun createTeams(teamRequest: TeamRequest, userProfileDto: UserProfileDto): TeamResponse {
+
         if(teamRepository.existsByName(teamRequest.name)) throw DuplicateArgumentException("Team", teamRequest.name)
-        //TODO("Team Repository.save 로 팀 생성")
         val teamResult = teamRepository.save(
             Team(
                 name = teamRequest.name,
             )
         )
-        //TODO("Team 생성 사용자 -> 리더로 권한 변경")
-        //TODO("이슈와 맴버의 개수를 세는 로직 작성")
-//        val leader = memberService.findById(memberId)
-//
-//        teamResult.id?.let { leader.giveTeamId(it) }
+        val leader = memberRepository.findByEmail(userProfileDto.email)
 
-        //TODO("TeamResponse 에 작성된 팀 정보 반환")
-        return TeamResponse.from(teamResult, teamResult.getIssuesSize(), teamResult.getMembersSize(), teamResult.members)
+        teamResult.let { leader?.giveTeam(it) }
+
+        return TeamResponse.from(teamResult, teamResult.getIssuesSize(), 1, mutableListOf(leader!!))
     }
 
     @Transactional(readOnly = true)
-    fun getTeamList(name: String, /*memberId : Long*/): List<TeamResponse> {
-        //TODO("authentication 에서 접근 사용자가 관리자 인지 확인")
+    fun getTeamList(name: String, userProfileDto: UserProfileDto): List<TeamResponse> {
+
+        if(userProfileDto.role == Role.ADMIN.name) throw NotAuthorityException("권한이 없습니다", userProfileDto.role)
+
         val teamResult = teamRepository.findAll()
         //TODO("name 에 특정 값이 들어올 경우 들어 온 값으로 Team Repository 에서 필터링 후에 조회")
 
@@ -48,22 +48,25 @@ class TeamService(
     }
 
     @Transactional(readOnly = true)
-    fun getTeamById(teamId: Long, /*memberId : Long*/): TeamResponse {
-        //TODO("authentication 에서 접근 사용자의 권한 확인")
-        
-        //TODO("권한이 사용자 와 리더 일 경우 teamId 가 authentication 에서 teamName 을 비교 후에 일치 하지 않으면 throw illegalArgumentException")
-        
-        //TODO("권한이 관리자일 경우 소속팀 여부와 상관 없이 모두 조회 가능")
+    fun getTeamById(teamId: Long, userProfileDto: UserProfileDto): TeamResponse {
+
+        if(userProfileDto.role == Role.USER.name || userProfileDto.role == Role.LEADER.name){
+            val member = memberRepository.findByEmail(userProfileDto.email)
+            if(member?.team?.id != teamId) throw IllegalArgumentException("다른 팀을 선택 하셨습니다 사용 권한이 업습니다")
+        }
 
         val teamResult = teamRepository.findByIdOrNull(teamId) ?: throw ModelNotFoundException("Team", teamId.toString())
 
         return TeamResponse.from(teamResult, teamResult.getIssuesSize(), teamResult.getMembersSize(), teamResult.members)
     }
 
-    fun updateTeamById(teamId: Long, teamRequest: TeamRequest, /*memberId : Long*/): TeamResponse {
-        //TODO("authentication 에서 접근 사용자의 권한 확인")
-        //TODO("권한이 사용자 와 리더 일 경우 teamId 가 authentication 에서 teamName 을 비교 후에 일치 하지 않으면 throw illegalArgumentException")
-        //TODO("권한이 관리자일 경우 소속팀 여부와 상관 없이 모두 조회 가능")
+    fun updateTeamById(teamId: Long, teamRequest: TeamRequest, userProfileDto: UserProfileDto): TeamResponse {
+
+        if(userProfileDto.role == Role.USER.name || userProfileDto.role == Role.LEADER.name){
+            val member = memberRepository.findByEmail(userProfileDto.email)
+            if(member?.team?.id != teamId) throw IllegalArgumentException("다른 팀을 선택 하셨습니다 사용 권한이 업습니다")
+        }
+
         val teamResult = teamRepository.findByIdOrNull(teamId) ?: throw ModelNotFoundException("Team", teamId.toString())
 
         teamResult.name = teamRequest.name
@@ -71,20 +74,29 @@ class TeamService(
         return TeamResponse.from(teamResult, teamResult.getIssuesSize(), teamResult.getMembersSize(), teamResult.members)
     }
 
-    fun deleteTeamById(teamId: Long, /*memberId : Long*/) {
-        //TODO("authentication 에서 접근 사용자의 권한 확인")
-        //TODO("권한이 리더 일 경우 teamId 가 authentication 에서 teamName 을 비교 후에 일치 하지 않으면 throw illegalArgumentException")
-        //TODO("권한이 관리자 일 경우 소속팀 여부와 상관 없이 모두 삭제 가능")
-        //TODO("권한이 사용자 일 경우 throw NotAuthenticatedException")
+    fun deleteTeamById(teamId: Long, userProfileDto: UserProfileDto) {
+        when(userProfileDto.role) {
+            Role.LEADER.name ->{
+                val member = memberRepository.findByEmail(userProfileDto.email)
+                if(member?.team?.id != teamId) throw IllegalArgumentException("다른 팀을 선택 하셨습니다 사용 권한이 없습니다")
+            }
+            Role.USER.name -> throw NotAuthorityException("사용 권한이 없습니다", userProfileDto.role)
+        }
+
+        if(teamId == 1L) throw IllegalArgumentException("기본 팀은 선택할 수 없습니다")
+
         val teamResult = teamRepository.findByIdOrNull(teamId) ?: throw ModelNotFoundException("Team", teamId.toString())
+        val dummyTeam = teamRepository.findByIdOrNull(1L)!!
+
+        teamResult.members.map{ it.team = dummyTeam }
 
         teamRepository.delete(teamResult)
     }
 
-    fun inviteMember(memberId: Long, /*memberId : Long*/): TeamResponse {
-        //TODO("authentication 에서 접근 사용자의 권한 확인")
-        //TODO("소속 팀의 리더가 아닐 경우 throw NotAuthenticatedException")
-        val leader = memberRepository.findByIdOrNull(6L) ?: throw ModelNotFoundException("Member", memberId.toString())
+    fun inviteMember(memberId: Long, userProfileDto: UserProfileDto): TeamResponse {
+        if(userProfileDto.role != Role.LEADER.name) throw NotAuthorityException("해당 맴버는 리더가 아닙니다", userProfileDto.role)
+
+        val leader = memberRepository.findByEmail(userProfileDto.email) ?: throw ModelNotFoundException("Member", memberId.toString())
         val member = memberRepository.findByIdOrNull(memberId) ?: throw ModelNotFoundException("Member", memberId.toString())
 
         when(member.team!!.id){
@@ -93,14 +105,13 @@ class TeamService(
             else -> throw IllegalArgumentException("해당 맴버는 다른 팀에 소속이 되어 있습니다")
         }
         val teamResult = teamRepository.findByIdOrNull(leader.team!!.id) ?: throw ModelNotFoundException("Team", leader.team!!.id.toString())
+
         return TeamResponse.from(teamResult, teamResult.getIssuesSize(), teamResult.getMembersSize(), teamResult.members)
     }
 
-    fun firedMember(memberId: Long, /*memberId : Long*/):TeamResponse{
-        //TODO("authentication 에서 접근 사용자의 권한 확인")
-        //TODO("소속 팀의 리더나 관리자 가 아닐 경우 throw NotAuthenticatedException")
-        //TODO("memberId 를 조회 시 없을 경우 throw ModelNotFoundException")
-        //TODO("memberId 의 정보를 가져 왔을 때 팀의 id가 authentication 에서 접근 사용자의 TeamId 와 다를 경우 throw illegalArgumentException")
+    fun firedMember(memberId: Long, userProfileDto: UserProfileDto):TeamResponse{
+        if(userProfileDto.role != Role.LEADER.name || userProfileDto.role != Role.ADMIN.name) throw NotAuthorityException("리더와 관리자만 접근이 가능 합니다", userProfileDto.role)
+
         val leader = memberRepository.findByIdOrNull(6L) ?: throw ModelNotFoundException("Member", memberId.toString())
         val member = memberRepository.findByIdOrNull(memberId) ?: throw ModelNotFoundException("Member", memberId.toString())
